code change for DraftConverter 1. rename DraftHtmlConverter to DraftConverter 2. DraftConverter.convertToHtml and DraftConverter.convertToApiData
